fix data types in clioinfra unpivot queries

+++ clioinfra/unpivot_gdp.sql
@@ -5,7 +5,7 @@ create extension if not exists hstore;
 
 CREATE table geoinfra.gdp_unpivot
 AS 
-SELECT year, (h).key as name, (h).value As amount
+SELECT year, (h).key as name, (h).value::integer As amount
  FROM (SELECT year, each(hstore(foo) - 'year'::text) As h
   FROM geoinfra.gdp as foo  ) As unpiv ;
